+++ src/Globals.c
@@ -11,8 +11,8 @@ uint8_t g_arr_tile_definitions_size = 0;
 
 // Game State
 TileState** g_game_board = NULL;
-int camera_offset_x = 0;
-int camera_offset_y = 0;
+float camera_offset_x = 0.0f;
+float camera_offset_y = 0.0f;
 
 void cleanup_globals(){
     free(g_arr_tile_definitions);
